drivers:video:video_stm32_dcmi support reuse buffer
There are times, for me the majority of times, that if the
camera is faster than the code/device using it, I would
rather keep the newest camera images versus the first
ones that were captured.

Currently the code, tries to grab a buffer from the
free pool and if none it throws away the new image.

This version optionally has a setting, that instead if
the free pool does not have any free items, it will
instead grab the first item from the out fifo, which
is the oldest item and use it instead.