video: stm32_dcmi:  DMA Error recovery
On several boards, such as the Arduino Giga and
Portenta H7, they are often times setup with their
camera buffers and potentially video buffers in
SDRam.  This can lead to a significant number of
DMA errors, which currently stops the camera from
returning any additional frames.

Fix the HAL_DCMI_Start_DMA call to ask for
the size of the image instead of the size of
the buffer

Signed-off-by: Kurt Eckhardt <kurte@rockisland.com>