video: stm32_dcmi: Snapshot mode
Add the ability to run cameras in a logical
snapshot mode instead of always running
in video mode.  In particular, the camera
starts when you ask for an image to be dequeued
and stops when it receives an image.

There are three ways to setup to run in this mode:

a) define a max of 1 buffer:
CONFIG_VIDEO_BUFFER_POOL_NUM_MAX=1

b) Tell it to do so in the overlay by giving the DCMI object
the property: snapshot-mode

c) At run time:  I added a video control: VIDEO_CID_SNAPSHOT_MODE
Which can have the values 0 - normal, 1 - snapshot

As mentioned in snapshot mode, the HAL DMA is stopped
after each image and restarted when the user asks for another
image.  I added a timeout default to 1 second, that handles
the case where sometimes the HAL is silently not successful
in retrieving an image.

Signed-off-by: Kurt Eckhardt <kurte@rockisland.com>