video: stm32_dcmi: snapshot stream stop
Experimenting with the optional snapshot_mode=2

With this it calls off to stop the video stream after each frame
is returned and restarts it when you ask for a new frame.

Testing to see if it helps or hurts the throughput speed.

Also testing if it helps/hurts with image tearing after
DMA errors.

Signed-off-by: Kurt Eckhardt <kurte@rockisland.com>