drivers: video: hm01b0 - reset/pwdn more formats
Adding some of the missing pieces to the HM01b0 video class.

Note: not maybe necessary, but changed the VIDEO define to match
all of the other video drivers that I have tried:
That is CONFIG_VIDEO_HIMAX_HM01B0 to
CONFIG_VIDEO_HM01B0

Added PWDN and RESET pins, to yaml, plus code.
Note: my handling of these two pins is similar to what OpenCV or
was it ME that if the first configuration of the pins
HIGH versus LOW), it tries all of the combinations until one
works or it completely fails.

Added a few more format sizes: 324x324 to 324x244 match the spec.
Note in my earlier runs I found I needed 326, so will test this out
more as it is what the spec says.

Signed-off-by: Kurt Eckhardt <kurte@rockisland.com>